feat(docker_modules/isaac_ros): Support Isaac ROS AprilTag
Fixes the VPI error:

```
user@ubuntu:~/workspaces/isaac_ros-dev/isaac_ros_assets$ ros2 launch isaac_ros_examples isaac_ros_examples.launch.py launch_fragments:=apriltag interface_specs_file:=${ISAAC_ROS_WS}/isaac_ros_assets/isaac_ros_apriltag/quickstart_interface_specs.json
[component_container_mt-1] [INFO] [isaac_ros_examples.container]: Load Library: /opt/ros/humble/lib/libapriltag_node.so
[component_container_mt-1] [ERROR] [isaac_ros_examples.container]: Failed to load library: Could not load library dlopen error: libnvvpi.so.3: cannot open shared object file: No such file or directory, at ./src/shared_library.c:99
[ERROR] [launch_ros.actions.load_composable_nodes]: Failed to load node 'apriltag' of type 'nvidia::isaac_ros::apriltag::AprilTagNode' in container '/isaac_ros_examples/container': Failed to load library: Could not load library dlopen error: libnvvpi.so.3: cannot open shared object file: No such file or directory, at ./src/shared_library.c:99
```

Ref: https://nvidia-isaac-ros.github.io/repositories_and_packages/isaac_ros_apriltag/isaac_ros_apriltag/index.html#quickstart

+++ docker_modules/install_isaac_ros.sh
@@ -24,14 +24,12 @@ if [ -z "$ISAAC_ROS_WS" ]; then
     echo "Error: ISAAC_ROS_WS environment variable is required but not set"
     exit 1
 fi
-# TODO: Check if this check is necessary
-# Temporarily ignore CUDA Toolkit version check
-# if [ "$CUDA_TOOLKIT_VERSION" != "12.6" ]; then
-#     # Ref: https://github.com/NVIDIA-ISAAC-ROS/isaac_ros_common/blob/main/docker/Dockerfile.base
-#     echo "Error: CUDA_TOOLKIT_VERSION must be 12.6 for Isaac ROS"
-#     # Other versions may be supported, but we haven't tested them yet.
-#     exit 1
-# fi
+if [ "$CUDA_TOOLKIT_VERSION" != "12.6" ]; then
+    # Ref: https://github.com/NVIDIA-ISAAC-ROS/isaac_ros_common/blob/main/docker/Dockerfile.base
+    echo "Error: CUDA_TOOLKIT_VERSION must be 12.6 for Isaac ROS"
+    # Other versions may be supported, but we haven't tested them yet.
+    exit 1
+fi
 
 echo "Installing Isaac ROS components for architecture: $TARGETARCH"
 echo "Isaac ROS: $ISAAC_ROS"
@@ -47,14 +45,45 @@ if [ "$ROS_DISTRO" != "humble" ]; then
     exit 1
 fi
 
+if [ "$TARGETARCH" == "arm64" ]; then
+    echo "Not Implemented: Isaac ROS installation for architecture: $TARGETARCH"
+    exit 1
+fi
+
 if [ "$ISAAC_ROS" = "YES" ]; then
     echo "Installing Isaac ROS..."
+
+    echo "Add Isaac apt repository..."
     # Ref: https://nvidia-isaac-ros.github.io/getting_started/isaac_apt_repository.html
     wget -qO - https://isaac.download.nvidia.com/isaac-ros/repos.key | sudo apt-key add -
     grep -qxF "deb https://isaac.download.nvidia.com/isaac-ros/release-3 $(lsb_release -cs) release-3.0" /etc/apt/sources.list || \
     echo "deb https://isaac.download.nvidia.com/isaac-ros/release-3 $(lsb_release -cs) release-3.0" | sudo tee -a /etc/apt/sources.list
     sudo apt-get update && sudo rm -rf /var/lib/apt/lists/* || exit 1
 
+    echo "Setup Jetson debian repositories..."
+    # Ref: https://github.com/NVIDIA-ISAAC-ROS/isaac_ros_common/blob/e516c452bda914b631dac626e457594b6b73c2ea/docker/Dockerfile.base#L215-L226
+    sudo apt-key adv --fetch-key https://repo.download.nvidia.com/jetson/jetson-ota-public.asc
+    if [[ ${TARGETARCH} == 'arm64' ]]; then
+        echo 'deb https://repo.download.nvidia.com/jetson/common r36.4 main' | sudo tee -a /etc/apt/sources.list.d/nvidia-l4t-apt-source.list
+        echo 'deb https://repo.download.nvidia.com/jetson/t234 r36.4 main' | sudo tee -a /etc/apt/sources.list.d/nvidia-l4t-apt-source.list
+    elif [[ ${TARGETARCH} == 'amd64' ]]; then
+        sudo add-apt-repository "deb http://repo.download.nvidia.com/jetson/x86_64/$(lsb_release -cs) r36.4 main"
+    else
+        echo "Unrecognized platform: ${TARGETARCH}"
+        exit 1
+    fi
+    sudo apt-get update && sudo rm -rf /var/lib/apt/lists/* || exit 1
+
+    echo "Install VPI..."
+    # Ref: https://github.com/NVIDIA-ISAAC-ROS/isaac_ros_common/blob/e516c452bda914b631dac626e457594b6b73c2ea/docker/Dockerfile.base#L240C1-L257C13
+    sudo apt-get update && sudo apt-get install -y \
+        libnvvpi3 \
+        vpi3-dev \
+        && sudo rm -rf /var/lib/apt/lists/* || exit 1
+    # TODO: Add VPI installation for arm64
+    # Ref: https://nvidia-isaac-ros.github.io/getting_started/hardware_setup/compute/jetson_vpi.html
+
+    echo "Cloning Isaac ROS common..."
     # Ref: https://nvidia-isaac-ros.github.io/repositories_and_packages/isaac_ros_nvblox/isaac_ros_nvblox/index.html
     mkdir -p ${ISAAC_ROS_WS}/src
     cd ${ISAAC_ROS_WS}/src
